// src/app/model/employee.model.ts
export class Employee {

    id: number | undefined;
    firstName: string | undefined;
    lastName: string | undefined;
    dateOfBirth: Date | undefined;
    phone: string | undefined;
    gender: string | undefined;
    skillName: string | undefined;
    experience: string | undefined;
    skillLevel: string | undefined;
    isValid: boolean | undefined;
    
}